Import Mesa 23.1.9

+++ lib/mesa/.gitlab-ci/container/build-deqp.sh
@@ -7,17 +7,25 @@ git config --global user.email "mesa@example.com"
 git config --global user.name "Mesa CI"
 git clone \
     https://github.com/KhronosGroup/VK-GL-CTS.git \
-    -b vulkan-cts-1.3.3.0 \
+    -b vulkan-cts-1.3.5.0 \
     --depth 1 \
     /VK-GL-CTS
 pushd /VK-GL-CTS
 
-# Apply a patch to update zlib link to an available version.
-# vulkan-cts-1.3.3.0 uses zlib 1.2.12 which was removed from zlib server due to
-# a CVE. See https://zlib.net/
-# FIXME: Remove this patch when uprev to 1.3.4.0+
-wget -O- https://github.com/KhronosGroup/VK-GL-CTS/commit/6bb2e7d64261bedb503947b1b251b1eeeb49be73.patch |
-    git am -
+cts_commits_to_backport=()
+
+for commit in "${cts_commits_to_backport[@]}"
+do
+  curl -L --retry 4 -f --retry-all-errors --retry-delay 60 \
+    "https://github.com/KhronosGroup/VK-GL-CTS/commit/$commit.patch" | git am -
+done
+
+# Fix surfaceless build.
+git am < $OLDPWD/.gitlab-ci/container/0001-Fix-build-for-the-surfaceless-and-null-WS-target-pla.patch
+
+# Android specific patches.
+git am < $OLDPWD/.gitlab-ci/container/0001-Allow-running-on-Android-from-the-command-line.patch
+git am < $OLDPWD/.gitlab-ci/container/0002-Android-prints-to-stdout-instead-of-logcat.patch
 
 # --insecure is due to SSL cert failures hitting sourceforge for zlib and
 # libpng (sigh).  The archives get their checksums checked anyway, and git
@@ -31,15 +39,24 @@ cp doc/testlog-stylesheet/testlog.{css,xsl} /deqp
 popd
 
 pushd /deqp
-# When including EGL/X11 testing, do that build first and save off its
-# deqp-egl binary.
-cmake -S /VK-GL-CTS -B . -G Ninja \
-      -DDEQP_TARGET=x11_egl_glx \
-      -DCMAKE_BUILD_TYPE=Release \
-      $EXTRA_CMAKE_ARGS
-ninja modules/egl/deqp-egl
-cp /deqp/modules/egl/deqp-egl /deqp/modules/egl/deqp-egl-x11
 
+if [ "${DEQP_TARGET}" != 'android' ]; then
+    # When including EGL/X11 testing, do that build first and save off its
+    # deqp-egl binary.
+    cmake -S /VK-GL-CTS -B . -G Ninja \
+        -DDEQP_TARGET=x11_egl_glx \
+        -DCMAKE_BUILD_TYPE=Release \
+        $EXTRA_CMAKE_ARGS
+    ninja modules/egl/deqp-egl
+    cp /deqp/modules/egl/deqp-egl /deqp/modules/egl/deqp-egl-x11
+
+    cmake -S /VK-GL-CTS -B . -G Ninja \
+        -DDEQP_TARGET=wayland \
+        -DCMAKE_BUILD_TYPE=Release \
+        $EXTRA_CMAKE_ARGS
+    ninja modules/egl/deqp-egl
+    cp /deqp/modules/egl/deqp-egl /deqp/modules/egl/deqp-egl-wayland
+fi
 
 cmake -S /VK-GL-CTS -B . -G Ninja \
       -DDEQP_TARGET=${DEQP_TARGET:-x11_glx} \
@@ -47,7 +64,9 @@ cmake -S /VK-GL-CTS -B . -G Ninja \
       $EXTRA_CMAKE_ARGS
 ninja
 
-mv /deqp/modules/egl/deqp-egl-x11 /deqp/modules/egl/deqp-egl
+if [ "${DEQP_TARGET}" != 'android' ]; then
+    mv /deqp/modules/egl/deqp-egl-x11 /deqp/modules/egl/deqp-egl
+fi
 
 # Copy out the mustpass lists we want.
 mkdir /deqp/mustpass
@@ -56,28 +75,30 @@ for mustpass in $(< /VK-GL-CTS/external/vulkancts/mustpass/main/vk-default.txt) 
         >> /deqp/mustpass/vk-master.txt
 done
 
-cp \
-    /deqp/external/openglcts/modules/gl_cts/data/mustpass/gles/aosp_mustpass/3.2.6.x/*.txt \
-    /deqp/mustpass/.
-cp \
-    /deqp/external/openglcts/modules/gl_cts/data/mustpass/egl/aosp_mustpass/3.2.6.x/egl-master.txt \
-    /deqp/mustpass/.
-cp \
-    /deqp/external/openglcts/modules/gl_cts/data/mustpass/gles/khronos_mustpass/3.2.6.x/*-master.txt \
-    /deqp/mustpass/.
-cp \
-    /deqp/external/openglcts/modules/gl_cts/data/mustpass/gl/khronos_mustpass/4.6.1.x/*-master.txt \
-    /deqp/mustpass/.
-cp \
-    /deqp/external/openglcts/modules/gl_cts/data/mustpass/gl/khronos_mustpass_single/4.6.1.x/*-single.txt \
-    /deqp/mustpass/.
+if [ "${DEQP_TARGET}" != 'android' ]; then
+    cp \
+        /deqp/external/openglcts/modules/gl_cts/data/mustpass/gles/aosp_mustpass/3.2.6.x/*.txt \
+        /deqp/mustpass/.
+    cp \
+        /deqp/external/openglcts/modules/gl_cts/data/mustpass/egl/aosp_mustpass/3.2.6.x/egl-master.txt \
+        /deqp/mustpass/.
+    cp \
+        /deqp/external/openglcts/modules/gl_cts/data/mustpass/gles/khronos_mustpass/3.2.6.x/*-master.txt \
+        /deqp/mustpass/.
+    cp \
+        /deqp/external/openglcts/modules/gl_cts/data/mustpass/gl/khronos_mustpass/4.6.1.x/*-master.txt \
+        /deqp/mustpass/.
+    cp \
+        /deqp/external/openglcts/modules/gl_cts/data/mustpass/gl/khronos_mustpass_single/4.6.1.x/*-single.txt \
+        /deqp/mustpass/.
 
-# Save *some* executor utils, but otherwise strip things down
-# to reduct deqp build size:
-mkdir /deqp/executor.save
-cp /deqp/executor/testlog-to-* /deqp/executor.save
-rm -rf /deqp/executor
-mv /deqp/executor.save /deqp/executor
+    # Save *some* executor utils, but otherwise strip things down
+    # to reduct deqp build size:
+    mkdir /deqp/executor.save
+    cp /deqp/executor/testlog-to-* /deqp/executor.save
+    rm -rf /deqp/executor
+    mv /deqp/executor.save /deqp/executor
+fi
 
 # Remove other mustpass files, since we saved off the ones we wanted to conventient locations above.
 rm -rf /deqp/external/openglcts/modules/gl_cts/data/mustpass
